// src/types/errors/errorHttp.ts
export enum ErrorMessage {
  FORBIDDEN = 'FORBIDDEN',
  BAD_REQUEST = 'BAD_REQUEST',
  UNKNOW_ERROR = 'UNKNOW_ERROR',
  JWT_EXPIRED = 'JWT_EXPIRED',
  NOT_UNIQ = 'NOT_UNIQ',
  BAD_CREDENTIALS = 'BAD_CREDENTIALS',
  NOT_SAME = 'NOT_SAME',
  CANT_PARTICIPATE = 'CANT_PARTICIPATE',
  POST_NOT_EXIST = 'POST_NOT_EXIST',
  COMMENT_NOT_EXIST = 'COMMENT_NOT_EXIST',
}

export enum ErrorDb {
  NOT_UNIQ = 'P2002',
}